Avoid race between interface & devices
This only affects startup & cases where a device and the instance this
device uses are created at the same time. Previously this would result
in a race between the interface creating the manager and the device
attempting to use it, which the device almost always won as it has way
less work to do before getting to the point where the two meet

Now instead of erroring out when a device attempts to use an instance
for which no manager exists it waits for some time for it to become
available. This will time out so cases where a device uses an instance
that does not exist will still error out, albeit slightly slower than
before.