Validate options & capabilities
This uses the definitions provided by Zigbee2MQTT on the
`bridge/devices` topic to validate the data in the CRD before sending it
to Zigbee2MQTT.

This results in better feedback as not all invalid data will result in a
response from Zigbee2MQTT that can be related back to the cause. It also
avoids putting a device in an invalid state and spamming the Zigbee2MQTT
log with errors.