Move TopicTracker locks from manager to TT itself
This allows the Manager to return TT instances, reducing the amount of
wrapper methods needed on it as it no longer needs to wrap every method
separately.

The downside is that the users need to so a bit more error handling as
they now need to consider both a failure to create the tracker as well
as as failure of the method they are calling, which previously were
combined in a single result.