Reworked event/error handling
We were doing a decent job of emitting useful events for errors, but our
approach had a few downsides causing it to be brittle and verbose,
mostly caused by the desire to attach a useful path to the event.

In many cases the place where an error occurs often knows what specific
value caused the error but not what the relevant path is, while the
place that supplied the value knows this path for all the values it has
supplied but doesn't know which of these values caused the error.

This led to a patchwork of passing paths and returning things with
paths, which was overly verbose and not neccesarily correct. In addition
it led to a lot of passing around of `EventManager` instances.

---

To resolve these issues this refactoring introduces a struct
`ValueWithSource`, which (as the name implies) is an arbitrary value
with information about the source of this value.

This means that this source information will be available when we create
an error, enabling us to attach the relevant source information to this
error.

These errors (with source information) can then propagate up and be
handled in a centralized location, avoiding passing `EventManager`
around just to handle errors.