Improve instance status management
The status was previously only updated in the status reporter task. This
meant that there were no updates before a connection was made to the
broker, so if this connection was never established this wouldn't be
reflected in the status. Similarly, if the status reporter task stopped
the connection would be terminated but this was not reflected in the
status.