Improve instance manager management in state
The previous setup allowed devices/groups to wait for an instance's
manager to become available with a simple timeout. This worked, but
because there was no understanding of the underlying CRD this was
somewhat crude; there was no difference between depending on an instance
that did not yet having a manager (which shouldn't be considered an
error) and depending on one that didn't exist (which should be
considered an error).

This has now been replaced with a manager that does understand the
instance CRD. When a resource attempts to get the manager for its
instance there are now 3 possible outcomes:

- The instance exists and has a manager: the manager is returned.
- The instance exists but does not yet have a manager: the resource's
  reconcile is rescheduled.
- The instance does not exist: the resource's reconcile fails with an
  appropriate error.