Build cross-compiled image on host platform
Rust can cross-compile, so the build itself doesn't have to be run on
the target platform. Running as host platform avoids qemu, _drastically_
speeding up the build.

As I write this the previous version has been running for over an hour
on Github and has yet to finish. The new version took 8 minutes on my
machine for the same platforms.