Fix option clearing with extension
Un-setting/clearing an uption has always been a pain. Previously this
could be done by passing in null, but this had the side effect of being
unable to set values to `null`, which was useful for some properties. In
the 2.0 version of Zigbee2MQTT this problem has been recognized and it
now stores `null` values (in some locations). That solves the issue of
not being able to set `null`, but it creates the new issue of not being
able to clear those properties.

To solve this the extension for this operator now adds a request type
that can be used to clear properties.