Use flake in github workflow
Previously the latest stable/nightly Rust toolchain was pulled in during
the workflow & used for the actions, but this results in the code being
checked with a different version of the tools than it was developed
against, which occasionally causes issues.

When all commits were done by a human this was just an annoyance, but
with automated updates this is now more of a problem, causing version
bumps to randomly fail because clippy has introduced a new check that
has nothing to do with the component being updated.

This commit resolves that by using the tools setup in `flake.nix` in the
workflows, ensuring that any issues causes by an update of the
toolchain/tooling will only be triggered by the commit that _actually_
updates those components.